+++ examples/helloworld/world/main.go
@@ -3,7 +3,7 @@
 
 package main
 
-// Nute is a basic gomobile app.
+// World is a basic gomobile app.
 import (
 	"flag"
 	"log"
@@ -30,11 +30,11 @@ func main() {
 	callerCreds := flag.String("CREDS", "", "Credentials of caller")
 	insecure := flag.Bool("insecure", false, "Skip server validation")
 	flag.Parse()
-	nuteLog, err := os.OpenFile("nute.log", os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
+	worldLog, err := os.OpenFile("world.log", os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
 	if err != nil {
 		log.Fatal(err)
 	}
-	log.SetOutput(nuteLog)
+	log.SetOutput(worldLog)
 
 	server.InitServer(*callerCreds, *insecure)
 	app.Main(func(a app.App) {
